+++ logger.go
@@ -6,43 +6,38 @@ import (
 )
 
 // Returns a logger with log.LstdFlags
-func NewVerbose(w io.Writer) L {
-	return &pAdapter{log.New(w, "", log.LstdFlags)}
+func New(w io.Writer) L {
+	return &lAdapter{log.New(w, "", log.LstdFlags)}
 }
 
-// Returns a logger with log.LstdFlags|log.Lshortfile
-func NewDebug(w io.Writer) L {
-	return &pAdapter{log.New(w, "", log.LstdFlags|log.Lshortfile)}
-}
+type Any interface{}
 
 type L interface {
 	Log(args ...interface{})
 	Logf(format string, args ...interface{})
 }
 
-type pAdapter struct {
-	*log.Logger
-}
+type lAdapter struct{ P }
 
-func (p *pAdapter) Log(args ...interface{}) {
-	p.Print(args...)
-}
+func (a *lAdapter) Log(args ...interface{})            { a.Print(args...) }
+func (a *lAdapter) Logf(f string, args ...interface{}) { a.Printf(f, args...) }
 
-func (p *pAdapter) Logf(format string, args ...interface{}) {
-	p.Printf(format, args...)
-}
+func Adapt(printer P) L { return &lAdapter{printer} }
 
 type P interface {
 	Print(args ...interface{})
 	Printf(format string, args ...interface{})
 }
 
-// Returns a silent logger
-func New() L {
-	return &silent{}
+// Returns a logger with log.LstdFlags|log.Lshortfile
+func NewDebug(w io.Writer) L {
+	return &lAdapter{log.New(w, "", log.LstdFlags|log.Lshortfile)}
 }
 
-type silent struct{}
+// Logger that outputs nothing
+const Silent silent = iota
 
-func (*silent) Log(args ...interface{})                 {}
-func (*silent) Logf(format string, args ...interface{}) {}
+type silent int
+
+func (silent) Log(args ...interface{})                 {}
+func (silent) Logf(format string, args ...interface{}) {}