+++ logger_test.go
@@ -2,6 +2,8 @@ package logger
 
 import (
 	"bytes"
+	"fmt"
+	"log"
 	"testing"
 
 	"github.com/gregoryv/asserter"
@@ -10,25 +12,26 @@ import (
 func Test_output_of_logger(t *testing.T) {
 	buf := bytes.NewBufferString("")
 	assert := asserter.New(t)
-	verbose := NewVerbose(buf)
-	debug := NewDebug(buf)
 	cases := []struct {
 		l L
 	}{
-		{verbose},
-		{debug},
+		{New(buf)},
+		{NewDebug(buf)},
+		{Adapt(log.New(buf, "", log.LstdFlags))},
 	}
-	for _, c := range cases {
-		c.l.Log("hello")
-		assert().Contains(buf.Bytes(), "hello")
-		c.l.Logf("%s is ok", "world")
-		assert().Contains(buf.Bytes(), "world is ok")
+	for i, c := range cases {
+		exp := fmt.Sprintf("hello %v", i)
+		c.l.Log(exp)
+		assert().Contains(buf.Bytes(), exp)
+		buf.Reset()
+		c.l.Logf("%s", exp)
+		assert().Contains(buf.Bytes(), exp)
 		buf.Reset()
 	}
 }
 
 func Test_silent(t *testing.T) {
-	l := New()
+	l := Silent
 	l.Log("nada")
 	l.Logf("%s", "nada")
 	//t.Fail()